refactor and export loading as it's own component
